Move iptables HostPortManager code into its own file.
Put all of the iptables-specific code from hostport.go and
hostport_manager.go into hostport_iptables.go (and all of the
iptables-specific tests into hostport_iptables_test.go), leaving just
the type definitions in hostport_manager.go and test case data in
hostport_manager_test.go.
